+++ m-DengYufan/zone/zone.cpp
@@ -4,26 +4,26 @@
 
 using namespace std;
 
-const int kMove[4][2] = {{1, 0}, {0, 1}, {-1, 0}, {0, -1}};
-int n, m, ans;
+const LL kMove[4][2] = {{1, 0}, {0, 1}, {-1, 0}, {0, -1}};
+LL n, m, ans;
 bool b[8][9], _b[8][9];
 
-void _S(int x, int y, bool c) {
+void _S(LL x, LL y, bool c) {
   if (x < 1 || x > n || y < 1 || y > m || b[x][y] != c || _b[x][y]) {
     return;
   }
   _b[x][y] = 1;
-  for (int i = 0; i <= 3; i++) {
+  for (LL i = 0; i <= 3; i++) {
     _S(x + kMove[i][0], y + kMove[i][1], c);
   }
 }
 
-void S(int x, int y) {
+void S(LL x, LL y) {
   if (x == n + 1 && y == 1) {
     bool flag = 0, _flag = 0, __flag = 0, ___flag = 0;
     fill(&_b[0][0], &_b[n][m] + 1, 0);
-    for (int i = 1; i <= n; i++) {
-      for (int j = 1; j <= m; j++) {
+    for (LL i = 1; i <= n; i++) {
+      for (LL j = 1; j <= m; j++) {
         if (!b[i][j] && !flag) {
           flag = 1;
           _S(i, j, 0);
